requirements.txt
streamlit>=1.32.0
plotly>=5.20.0
pandas>=2.2.0
anthropic>=0.78.0
apscheduler>=3.10.4
pytz>=2024.1
python-dotenv>=1.0.0
requests>=2.31.0
